Wallah plein de changements
- integration du code de chauvin : rythme cardiaque ok, BP bof, arythmie bof
- min/max/average ok
- remise a zero quand on change de jour
- affichage valeurs instantanées

+++ MonitoringApp/app/src/main/java/com/example/android/monitoringapp/MonthSummaryActivity.java
@@ -163,9 +163,9 @@ public class MonthSummaryActivity extends AppCompatActivity {
         salinityMax =(TextView) findViewById(R.id.value_max_sodium_chloride);
         salinityMean =(TextView) findViewById(R.id.value_mean_sodium_chloride);
 
-        salinityMin.setText(Integer.toString(dataMonth.getMinimum_sodium_chloride()));
-        salinityMax.setText(Integer.toString(dataMonth.getMaximum_sodium_chloride()));
-        salinityMean.setText(Integer.toString(dataMonth.getAverage_sodium_chloride()));
+        salinityMin.setText(Double.toString(dataMonth.getMinimum_sodium_chloride()));
+        salinityMax.setText(Double.toString(dataMonth.getMaximum_sodium_chloride()));
+        salinityMean.setText(Double.toString(dataMonth.getAverage_sodium_chloride()));
 
     }
